Refactor: Move everything to library
